Renamed `linear_algebra` to `matrix`
