+++ CMakeLists.txt
@@ -8,4 +8,4 @@ set(SOURCE_FILES "main.cpp" "linear_algebra.h")
 #set(CMAKE_VERBOSE_MAKEFILE ON)
 
 # add_library(linear_algebra INTERFACE)
-add_executable(NeuralNetworks main.cpp src/neural_layer.cpp src/neural_layer.h src/neural_matrix.cpp src/neural_matrix.h src/linear_algebra.h src/neural_network.cpp src/neural_network.h src/linear_algebra.cpp)
+add_executable(NeuralNetworks main.cpp src/neural_layer.cpp src/neural_layer.h src/neural_matrix.cpp src/neural_matrix.h src/matrix.h src/neural_network.cpp src/neural_network.h src/matrix.cpp)
